python_impl_generic/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_kmp_llvm
absl-py=2.1.0=py311h06a4308_0
blas=1.0=openblas
bzip2=1.0.8=h4bc722e_7
c-ares=1.33.1=heb4867d_0
ca-certificates=2024.7.4=hbcca054_0
cuda-cudart=12.4.127=0
cuda-cupti=12.4.127=0
cuda-libraries=12.4.0=0
cuda-nvrtc=12.4.127=0
cuda-nvtx=12.4.127=0
cuda-opencl=12.6.37=0
cuda-runtime=12.4.0=0
cuda-version=12.6=3
cudnn=8.9.7.29=h092f7fd_3
filelock=3.13.1=py311h06a4308_0
fsspec=2024.6.1=py311h06a4308_0
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py311hc9b5ff0_0
grpcio=1.62.2=py311ha6695c7_0
jinja2=3.1.4=py311h06a4308_0
ld_impl_linux-64=2.40=hf3520f5_7
libabseil=20240116.2=cxx17_h6a678d5_0
libblas=3.9.0=16_linux64_openblas
libcblas=3.9.0=16_linux64_openblas
libcublas=12.4.2.65=0
libcufft=11.2.0.44=0
libcufile=1.11.0.15=0
libcurand=10.3.7.37=0
libcusolver=11.6.0.99=0
libcusparse=12.3.0.142=0
libexpat=2.6.2=h59595ed_0
libffi=3.4.2=h7f98852_5
libgcc=14.1.0=h77fa898_1
libgcc-ng=14.1.0=h69a702a_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=14.1.0=h77fa898_1
libgrpc=1.62.2=h15f2491_0
liblapack=3.9.0=16_linux64_openblas
libllvm14=14.0.6=hcd5def8_4
libmagma=2.8.0=h0af6554_0
libmagma_sparse=2.8.0=h0af6554_0
libnpp=12.2.5.2=0
libnsl=2.0.1=hd590300_0
libnvfatbin=12.6.20=0
libnvjitlink=12.4.99=0
libnvjpeg=12.3.1.89=0
libopenblas=0.3.21=h043d6bf_0
libprotobuf=4.25.3=h08a7969_0
libre2-11=2023.09.01=h5a48ba9_2
libsqlite=3.46.0=hde9e2c9_0
libstdcxx=14.1.0=hc0a3c3a_1
libstdcxx-ng=14.1.0=h4852527_1
libtorch=2.4.0=cuda120_h2b0da52_301
libuuid=2.38.1=h0b41bf4_0
libuv=1.48.0=h5eee18b_0
libxcrypt=4.4.36=hd590300_1
libzlib=1.3.1=h4ab18f5_1
llvm-openmp=18.1.8=hf5423f3_1
llvmlite=0.43.0=py311h6a678d5_0
markdown=3.4.1=py311h06a4308_0
markupsafe=2.1.3=py311h5eee18b_0
mkl=2023.2.0=h84fe81f_50496
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py311h06a4308_0
nccl=2.22.3.1=hbc370b7_1
ncurses=6.5=he02047a_1
networkx=3.3=py311h06a4308_0
numba=0.60.0=py311h4bc866e_0
numpy=1.26.4=py311h24aa872_0
numpy-base=1.26.4=py311hbfb1bba_0
openssl=3.3.1=hb9d3cd8_3
pip=24.2=pyh8b19718_1
protobuf=4.25.3=py311h12ddb61_0
pyqt5=5.15.11=pypi_0
pyqt5-qt5=5.15.14=pypi_0
pyqt5-sip=12.15.0=pypi_0
python=3.11.9=hb806964_0_cpython
python_abi=3.11=5_cp311
pytorch=2.4.0=cuda120_py311hf6aebf0_301
pytorch-cuda=12.4=hc786d27_6
re2=2023.09.01=h7f4b329_2
readline=8.2=h8228510_1
setuptools=72.2.0=pyhd8ed1ab_0
six=1.16.0=pyhd3eb1b0_1
sleef=3.6.1=h1b44611_3
sympy=1.13.2=pypyh2585a3b_103
tbb=2021.8.0=hdb19cb5_0
tensorboard=2.17.0=py311h06a4308_0
tensorboard-data-server=0.7.0=py311h52d8a92_1
tk=8.6.13=noxft_h4845f30_101
typing_extensions=4.11.0=py311h06a4308_0
tzdata=2024a=h8827d51_1
werkzeug=3.0.3=py311h06a4308_0
wheel=0.44.0=pyhd8ed1ab_0
xz=5.2.6=h166bdaf_0
zstd=1.5.6=ha6fb4c9_0
